requirements.txt
Django==1.9.2
Pillow==3.3.1
